Pull out graph utils from CodeGraph
